temple-identification-backend/src/requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
pymongo==4.2.0
tensorflow==2.16.1
Pillow==10.1
gunicorn==20.1.0
Werkzeug==2.2.2